basic_func12

+++ basic_func12.py
@@ -10,4 +10,4 @@ def main(a):
     X = a
     return X
 y = main(8/3)
-print(round(y,2))
+print(float(round(y,2)))
